update API delete category

+++ src/services/category.service.js
@@ -2,6 +2,7 @@
 
 const { BadRequestError } = require("../core/error-response");
 const { findByName, createCategory, getAllCategory, deleteCategory, updateCategoryById } = require("../models/repositories/category");
+const { getNumberOfProductsByCategory, getAllProductsByUser } = require("../models/repositories/product");
 const { removeUndefinedObject } = require("../utils");
 
 class CategoryService {
@@ -19,6 +20,10 @@ class CategoryService {
     }
 
     static deleteCategory = async (id) => {
+        const foundProducts = await getAllProductsByUser({filter: {
+            product_categories: {$in: [id]},
+        }, unSelect: ["product_categories", "__v"]});
+        if (foundProducts.length > 0) throw new BadRequestError('Category has products');
         return await deleteCategory(id);
     }
     static updateCategory = async (id, data) => {
